toasty::schema

Struct Deferred

pub struct Deferred<T> { /* private fields */ }

A field whose value is not loaded by default.

Deferred<T> wraps a T whose underlying column is excluded from default queries. After a normal load the value is unloaded and accessing it via get panics. Fetch the value with the per-field .exec() accessor on the record.

Deferred<Option<T>> is supported when the column is nullable.

Implementations

impl<T> Deferred<T>

pub fn is_unloaded(&self) -> bool

Returns true if the field has not been loaded.

pub fn get(&self) -> &T

Returns a reference to the loaded value.

Panics

Panics if the field has not been loaded.

pub fn unload(&mut self)

Clears the loaded value, returning the field to the unloaded state.

pub fn into_inner(self) -> T

Consumes this Deferred<T> and returns the loaded value.

Panics

Panics if the field has not been loaded.